+++ app/src/main/java/org/simple/clinic/questionnaireresponse/sync/QuestionnaireResponsePayload.kt
@@ -24,10 +24,10 @@ data class QuestionnaireResponsePayload(
     val facilityId: UUID,
 
     @Json(name = "last_updated_by_user_id")
-    val lastUpdatedByUserId: UUID,
+    val lastUpdatedByUserId: UUID?,
 
     @Json(name = "content")
-    val content: Map<String, String>,
+    val content: Map<String, Any>,
 
     @Json(name = "created_at")
     val createdAt: Instant,